refactor(utils): extract gemtext parsing into gemtext.py
Move the gemtext line/document helpers (_parse_heading/_parse_link/.../
parse_gemtext) into a new gemtext.py that depends only on the gemtext models.
utils.py re-exports parse_gemtext; the gemini response processing still in utils
resolves it via that re-import. No behavior change.
